+++ ignite/services/network/launch.go
@@ -64,7 +64,7 @@ func (n Network) TriggerLaunch(ctx context.Context, launchID uint64, launchTime 
 
 	msg := launchtypes.NewMsgTriggerLaunch(address, launchID, launchTime)
 	n.ev.Send(events.New(events.StatusOngoing, "Setting launch time"))
-	res, err := n.cosmos.BroadcastTx(n.account, msg)
+	res, err := n.cosmos.BroadcastTx(ctx, n.account, msg)
 	if err != nil {
 		return err
 	}
@@ -81,7 +81,7 @@ func (n Network) TriggerLaunch(ctx context.Context, launchID uint64, launchTime 
 }
 
 // RevertLaunch reverts a launched chain as a coordinator
-func (n Network) RevertLaunch(launchID uint64, chain Chain) error {
+func (n Network) RevertLaunch(ctx context.Context, launchID uint64, chain Chain) error {
 	n.ev.Send(events.New(events.StatusOngoing, fmt.Sprintf("Reverting launched chain %d", launchID)))
 
 	address, err := n.account.Address(networktypes.SPN)
@@ -90,7 +90,7 @@ func (n Network) RevertLaunch(launchID uint64, chain Chain) error {
 	}
 
 	msg := launchtypes.NewMsgRevertLaunch(address, launchID)
-	_, err = n.cosmos.BroadcastTx(n.account, msg)
+	_, err = n.cosmos.BroadcastTx(ctx, n.account, msg)
 	if err != nil {
 		return err
 	}