+++ ignite/services/network/launch_test.go
@@ -44,6 +44,7 @@ func TestTriggerLaunch(t *testing.T) {
 			Once()
 		suite.CosmosClientMock.
 			On("BroadcastTx",
+				context.Background(),
 				account,
 				&launchtypes.MsgTriggerLaunch{
 					Coordinator: addr,
@@ -144,6 +145,7 @@ func TestTriggerLaunch(t *testing.T) {
 			Once()
 		suite.CosmosClientMock.
 			On("BroadcastTx",
+				context.Background(),
 				account,
 				&launchtypes.MsgTriggerLaunch{
 					Coordinator: addr,
@@ -183,6 +185,7 @@ func TestTriggerLaunch(t *testing.T) {
 			Once()
 		suite.CosmosClientMock.
 			On("BroadcastTx",
+				context.Background(),
 				account,
 				&launchtypes.MsgTriggerLaunch{
 					Coordinator: addr,
@@ -238,6 +241,7 @@ func TestRevertLaunch(t *testing.T) {
 		suite.ChainMock.On("ResetGenesisTime").Return(nil).Once()
 		suite.CosmosClientMock.
 			On("BroadcastTx",
+				context.Background(),
 				account,
 				&launchtypes.MsgRevertLaunch{
 					Coordinator: addr,
@@ -246,7 +250,7 @@ func TestRevertLaunch(t *testing.T) {
 			Return(testutil.NewResponse(&launchtypes.MsgRevertLaunchResponse{}), nil).
 			Once()
 
-		revertError := network.RevertLaunch(testutil.LaunchID, suite.ChainMock)
+		revertError := network.RevertLaunch(context.Background(), testutil.LaunchID, suite.ChainMock)
 		require.NoError(t, revertError)
 		suite.AssertAllMocks(t)
 	})
@@ -263,6 +267,7 @@ func TestRevertLaunch(t *testing.T) {
 
 		suite.CosmosClientMock.
 			On("BroadcastTx",
+				context.Background(),
 				account,
 				&launchtypes.MsgRevertLaunch{
 					Coordinator: addr,
@@ -274,7 +279,7 @@ func TestRevertLaunch(t *testing.T) {
 			).
 			Once()
 
-		revertError := network.RevertLaunch(testutil.LaunchID, suite.ChainMock)
+		revertError := network.RevertLaunch(context.Background(), testutil.LaunchID, suite.ChainMock)
 		require.Error(t, revertError)
 		require.Equal(t, expectedError, revertError)
 		suite.AssertAllMocks(t)
@@ -296,6 +301,7 @@ func TestRevertLaunch(t *testing.T) {
 			Once()
 		suite.CosmosClientMock.
 			On("BroadcastTx",
+				context.Background(),
 				account,
 				&launchtypes.MsgRevertLaunch{
 					Coordinator: addr,
@@ -304,7 +310,7 @@ func TestRevertLaunch(t *testing.T) {
 			Return(testutil.NewResponse(&launchtypes.MsgRevertLaunchResponse{}), nil).
 			Once()
 
-		revertError := network.RevertLaunch(testutil.LaunchID, suite.ChainMock)
+		revertError := network.RevertLaunch(context.Background(), testutil.LaunchID, suite.ChainMock)
 		require.Error(t, revertError)
 		require.Equal(t, expectedError, revertError)
 		suite.AssertAllMocks(t)